backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.10.0
pydantic-settings==2.5.2
pyjwt==2.9.0
pytest==8.3.3
httpx==0.27.2
